Thinned the bottom of the objective clip dovetail
I'd inadvertently added 0.5+d to the thickness of the dovetail clip for
the objective.  That made it hard to clip in, and caused trouble when it
broke, as it often did.  The new, thinner bottom should be stretchier
and easier to snip off if desired.  Need to recompile the STL


Former-commit-id: d24e4cb2d818f6d2939ec7c31072d783ede38d02 [formerly 08bdb968db6f8ca238cc5b7fff314c61c924a456]
Former-commit-id: 92791c45440b9167f8f45c9cfe4e8433fdeda35e
Former-commit-id: fa2cd80f376f3794206fa6e1bf5f037bd2217870