Re-improved the band insertion process.
This must have been lost somewhere as I'm sure I did it before.  Maybe
it's lurking un-synced on my mac...  I've strengthened the elastic band
hooks, and made the insertion tool a bit deeper.  I've also widened the
gap in the feet to make it easier to put the bands in, and slimmed down
the shaft on the insertion tool.


Former-commit-id: 53ec1e06a3e575a71d6e5a8f6a1b47fab1c0da5f [formerly 8ae5f7218d5d3dc23c35d554176af4181efaa46e]
Former-commit-id: 2800b0cd52bd5df9159336cdf915d2f74a999da2
Former-commit-id: c5ae5b193798d4c8e75d71e73d5138c9b23c692e