Corrected lens-to-sensor distance in RMS optics
I did some testing and noticed that the distance from the lens to the
sensor was wrong in the RMS optics module.  This will partly be because
the Pi camera sensor is now no longer in the plane of the board (it's
raised up by the foam pad under the camera).  I've added a function to
the camera files that returns how high the sensor is, to allow it to be
correct for different cameras.


Former-commit-id: bb1e0c9421d49ac3564c318d07fe719cad4fb5be [formerly cc71a244f8208b14fcb412afe0fb46782deb817a]
Former-commit-id: 06eff74ca5aafecf502a3e080a2dcc5720982c03
Former-commit-id: 622519c9964e798cb4fdc60c5c7ae01887f570d4