Added condenser mount and moved access hole
Refactored a bunch of stuff to tidy it up, and added a mount for the condenser above the Z axis mechanism.  The access hole for the screw that mounts the objective is now to the side of, rather than above, the Z gear, which should make it much easier to tweak.


Former-commit-id: 514e7a9005f89a787586c4830d75decf8fd6242d [formerly 37bd1a95abc4c9baa80c76325e3e8e6ec3a292b9] [formerly c84c78eae3d66f9c5a01bad685bc5a22e75e14f1 [formerly b6d0fbde362c689f4d851634e46dbfa13935939f]]
Former-commit-id: 6ddb89fe45c0c6fdf6f1c73be92da57318279dc9 [formerly 1d4d235139bc36a92f1bb817d93ecbbf731f01be]
Former-commit-id: a8cceb6da9581328c3842d4bc7e2503879755fe6
Former-commit-id: 01ad8d48bf6da68455c24618853cf4fc8cfbf492